# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Your API",
      "description": "API documentation for your backend",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      }
    ],
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "role": {
              "type": "string",
              "enum": ["user", "admin"]
            }
          },
          "required": ["username", "email", "password"]
        },
        "Event": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "category": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "date": {
              "type": "string",
              "format": "date"
            },
            "location": {
              "type": "string"
            },
            "objective": {
              "type": "object",
              "properties": {
                "time": {
                  "type": "number"
                },
                "quantity": {
                  "type": "number"
                },
                "surface": {
                  "type": "number"
                }
              },
              "required": ["time", "quantity", "surface"]
            },
            "createdBy": {
              "type": "string",
              "format": "uuid"
            },
            "status": {
              "type": "string",
              "enum": ["En cours", "Terminé", "À venir"]
            },
            "participants": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            },
            "currentStepIndex": {
              "type": "number"
            },
            "steps": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "stepNumber": {
                    "type": "number"
                  },
                  "stepName": {
                    "type": "string"
                  }
                }
              }
            },
            "teamA": {
              "type": "object",
              "properties": {
                "participants": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "quantity": {
                  "type": "number"
                },
                "lightStatus": {
                  "type": "string",
                  "enum": ["red", "green"]
                }
              }
            },
            "teamB": {
              "type": "object",
              "properties": {
                "participants": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "quantity": {
                  "type": "number"
                },
                "lightStatus": {
                  "type": "string",
                  "enum": ["red", "green"]
                }
              }
            }
          },
          "required": ["title", "category", "description", "date", "location", "objective", "createdBy"]
        },
        "Article": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "text": {
              "type": "string"
            },
            "owner": {
              "type": "string",
              "format": "uuid"
            },
            "photo": {
              "type": "string"
            },
            "date": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": ["title", "text", "owner"]
        }
      }
    },
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Register a new user",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created"
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login a user",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User logged in"
            },
            "401": {
              "description": "Invalid credentials"
            }
          }
        }
      },
      "/events": {
        "get": {
          "summary": "Get all events",
          "responses": {
            "200": {
              "description": "A list of events",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new event",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Event created"
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/events/{id}": {
        "get": {
          "summary": "Get event by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Event details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            },
            "404": {
              "description": "Event not found"
            }
          }
        },
        "put": {
          "summary": "Update event by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Event updated"
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "Event not found"
            }
          }
        },
        "delete": {
          "summary": "Delete event by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Event deleted"
            },
            "404": {
              "description": "Event not found"
            }
          }
        }
      },
      "/articles": {
        "get": {
          "summary": "Get all articles",
          "responses": {
            "200": {
              "description": "A list of articles",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Article"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new article",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Article created"
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/articles/{id}": {
        "get": {
          "summary": "Get article by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Article details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            },
            "404": {
              "description": "Article not found"
            }
          }
        },
        "put": {
          "summary": "Update article by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Article updated"
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "Article not found"
            }
          }
        },
        "delete": {
          "summary": "Delete article by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Article deleted"
            },
            "404": {
              "description": "Article not found"
            }
          }
        }
      }
    }
  }
  